+++ app/blog/internal/biz/user_token.go
@@ -6,10 +6,10 @@ import (
 )
 
 type UserTokenRepo interface {
-	GetToken(ctx context.Context, userId uint64) string
+	GetToken(ctx context.Context, userId uint32) string
 	ValidateToken(ctx context.Context, token string) error
 	GenerateToken(ctx context.Context, user *v1.User) (string, error)
-	RemoveToken(ctx context.Context, userId uint64) error
+	RemoveToken(ctx context.Context, userId uint32) error
 }
 
 type UserTokenUseCase struct {
@@ -26,7 +26,7 @@ func (uc *UserTokenUseCase) GenerateToken(ctx context.Context, user *v1.User) (s
 	return uc.repo.GenerateToken(ctx, user)
 }
 
-func (uc *UserTokenUseCase) GetToken(ctx context.Context, userId uint64) string {
+func (uc *UserTokenUseCase) GetToken(ctx context.Context, userId uint32) string {
 	return uc.repo.GetToken(ctx, userId)
 }
 
@@ -34,6 +34,6 @@ func (uc *UserTokenUseCase) ValidateToken(ctx context.Context, token string) err
 	return uc.repo.ValidateToken(ctx, token)
 }
 
-func (uc *UserTokenUseCase) RemoveToken(ctx context.Context, userId uint64) error {
+func (uc *UserTokenUseCase) RemoveToken(ctx context.Context, userId uint32) error {
 	return uc.repo.RemoveToken(ctx, userId)
 }
